Fix handling of replace statements
There were 2 issues:

First, dependencies that are replaced should be removed from the list to
download. The replaced dependencies will never be used as they are
replaced. This apply to both local and version controlled replacements.

Second, the replacement dependencies should only be added to the list to
download if they are effectively replacing a dependencies. In other
words, replace statement replacing a module not already in the list of
dependencies should just be ignored as it will never be used.

Issue https://github.com/jfrog/jfrog-cli/issues/480